feat(backup): Merge backup refactoring
(cherry picked from commit 14ab2db8804e88555bc913e2dd1bda462d33a629)
