+++ worker/backup_handler.go
@@ -13,13 +13,23 @@
 package worker
 
 import (
+	"bytes"
+	"encoding/json"
 	"fmt"
 	"io"
+	"io/ioutil"
 	"net/url"
+	"os"
+	"path/filepath"
 	"sort"
+	"strings"
+	"time"
 
 	"github.com/dgraph-io/dgraph/protos/pb"
 	"github.com/dgraph-io/dgraph/x"
+	"github.com/golang/glog"
+	"github.com/minio/minio-go/v6"
+	"github.com/minio/minio-go/v6/pkg/credentials"
 
 	"github.com/pkg/errors"
 )
@@ -53,84 +63,225 @@ const (
 	tmpManifest = `manifest_tmp.json`
 )
 
-// UriHandler interface is implemented by URI scheme handlers.
-// When adding new scheme handles, for example 'azure://', an object will implement
-// this interface to supply Dgraph with a way to create or load backup files into DB.
-// For all methods below, the URL object is parsed as described in `newHandler' and
-// the Processor object has the DB, estimated tablets size, and backup parameters.
-type UriHandler interface {
-	// Handlers must know how to Write to their URI location.
-	// These function calls are used by both Create and Load.
-	io.WriteCloser
+// getConsolidatedManifest walks over all the backup directories and generates a master manifest.
+func getConsolidatedManifest(h UriHandler, uri *url.URL) (*MasterManifest, error) {
+	// If there is a master manifest already, we just return it.
+	if h.FileExists(backupManifest) {
+		manifest, err := readMasterManifest(h, backupManifest)
+		if err != nil {
+			return &MasterManifest{}, errors.Wrap(err, "Failed to read master manifest")
+		}
+		return manifest, nil
+	}
 
-	// GetManifest returns the master manifest, containing information about all the
-	// backups. If the backup directory is using old formats (version < 21.03) of manifests,
-	// then it will return a consolidated master manifest.
-	GetManifest(*url.URL) (*MasterManifest, error)
+	// Otherwise, we create a master manifest by going through all the backup directories.
+	paths := h.ListPaths("")
 
-	// GetManifests returns the list of manifest for the given backup series ID
-	// and backup number at the specified location. If backupNum is set to zero,
-	// all the manifests for the backup series will be returned. If it's greater
-	// than zero, manifests from one to backupNum will be returned.
-	GetManifests(*url.URL, string, uint64) ([]*Manifest, error)
+	var manifestPaths []string
+	suffix := filepath.Join(string(filepath.Separator), backupManifest)
+	for _, p := range paths {
+		if strings.HasSuffix(p, suffix) {
+			manifestPaths = append(manifestPaths, p)
+		}
+	}
 
-	// GetLatestManifest reads the manifests at the given URL and returns the
-	// latest manifest.
-	GetLatestManifest(*url.URL) (*Manifest, error)
+	sort.Strings(manifestPaths)
+	var mlist []*Manifest
 
-	// CreateBackupFile prepares the object or file to save the backup file.
-	CreateBackupFile(*url.URL, *pb.BackupRequest) error
-
-	// CreateManifest creates the given manifest.
-	CreateManifest(*url.URL, *MasterManifest) error
-
-	// Load will scan location URI for backup files, then load them via loadFn.
-	// It optionally takes the name of the last directory to consider. Any backup directories
-	// created after will be ignored.
-	// Objects implementing this function will be used for retrieving (dowload) backup files
-	// and loading the data into a DB. The restore CLI command uses this call.
-	Load(*url.URL, string, uint64, loadFn) LoadResult
-
-	// Verify checks that the specified backup can be restored to a cluster with the
-	// given groups. The last manifest of that backup should have the same number of
-	// groups as given list of groups.
-	Verify(*url.URL, *pb.RestoreRequest, []uint32) error
+	for _, path := range manifestPaths {
+		path = filepath.Dir(path)
+		_, path = filepath.Split(path)
+		m, err := readManifest(h, filepath.Join(path, backupManifest))
+		if err != nil {
+			return nil, errors.Wrap(err, "While Getting latest manifest")
+		}
+		m.Path = path
+		mlist = append(mlist, m)
+	}
+	return &MasterManifest{Manifests: mlist}, nil
 }
 
-// NewUriHandler parses the requested URI and finds the corresponding UriHandler.
-// If the passed credentials are not nil, they will be used to override the
-// default credentials (only for backups to minio or S3).
-// Target URI formats:
-//   [scheme]://[host]/[path]?[args]
-//   [scheme]:///[path]?[args]
-//   /[path]?[args] (only for local or NFS)
-//
-// Target URI parts:
-//   scheme - service handler, one of: "file", "s3", "minio"
-//     host - remote address. ex: "dgraph.s3.amazonaws.com"
-//     path - directory, bucket or container at target. ex: "/dgraph/backups/"
-//     args - specific arguments that are ok to appear in logs.
-//
-// Global args (if supported by the handler):
-//     secure - true|false turn on/off TLS.
-//      trace - true|false turn on/off HTTP tracing.
-//   compress - true|false turn on/off data compression.
-//    encrypt - true|false turn on/off data encryption.
-//
-// Examples:
-//   s3://dgraph.s3.amazonaws.com/dgraph/backups?secure=true
-//   minio://localhost:9000/dgraph?secure=true
-//   file:///tmp/dgraph/backups
-//   /tmp/dgraph/backups?compress=gzip
-func NewUriHandler(uri *url.URL, creds *x.MinioCredentials) (UriHandler, error) {
-	switch uri.Scheme {
-	case "file", "":
-		return &fileHandler{}, nil
-	case "minio", "s3":
-		return NewS3Handler(uri, creds)
+func readManifest(h UriHandler, path string) (*Manifest, error) {
+	var m Manifest
+	b, err := h.Read(path)
+	if err != nil {
+		return &m, errors.Wrap(err, "readManifest failed to read the file: ")
 	}
-	return nil, errors.Errorf("Unable to handle url: %s", uri)
+	if err := json.Unmarshal(b, &m); err != nil {
+		return &m, errors.Wrap(err, "readManifest failed to unmarshal: ")
+	}
+	return &m, nil
+}
 
+func readMasterManifest(h UriHandler, path string) (*MasterManifest, error) {
+	var m MasterManifest
+	b, err := h.Read(path)
+	if err != nil {
+		return &m, errors.Wrap(err, "readMasterManifest failed to read the file: ")
+	}
+	if err := json.Unmarshal(b, &m); err != nil {
+		return &m, errors.Wrap(err, "readMasterManifest failed to unmarshal: ")
+	}
+	return &m, nil
+}
+
+func getLatestManifest(h UriHandler, uri *url.URL) (*Manifest, error) {
+	if !h.DirExists("./") {
+		return &Manifest{}, errors.Errorf("getLatestManifest: The uri path: %q doesn't exists",
+			uri.Path)
+	}
+	manifest, err := getConsolidatedManifest(h, uri)
+	if err != nil {
+		return nil, errors.Wrap(err, "Get latest manifest failed while consolidation: ")
+	}
+	if len(manifest.Manifests) == 0 {
+		return &Manifest{}, nil
+	}
+	return manifest.Manifests[len(manifest.Manifests)-1], nil
+}
+
+func getManifest(h UriHandler, uri *url.URL) (*MasterManifest, error) {
+	if !h.DirExists("") {
+		return &MasterManifest{}, errors.Errorf("getManifest: The uri path: %q doesn't exists",
+			uri.Path)
+	}
+	manifest, err := getConsolidatedManifest(h, uri)
+	if err != nil {
+		return manifest, errors.Wrap(err, "Failed to get consolidated manifest: ")
+	}
+	return manifest, nil
+}
+
+func createManifest(h UriHandler, uri *url.URL, manifest *MasterManifest) error {
+	var err error
+	if !h.DirExists("./") {
+		if err := h.CreateDir("./"); err != nil {
+			return errors.Wrap(err, "createManifest failed to create path")
+		}
+	}
+
+	w, err := h.CreateFile(tmpManifest)
+	if err != nil {
+		return errors.Wrap(err, "createManifest failed to create tmp path")
+	}
+	if err = json.NewEncoder(w).Encode(manifest); err != nil {
+		return err
+	}
+	if err := w.Close(); err != nil {
+		return err
+	}
+	// Move the tmpManifest to backupManifest, this operation is not atomic for s3.
+	// We try our best to move the file but if it fails then the user must move it manually.
+	err = h.Rename(tmpManifest, backupManifest)
+	return errors.Wrapf(err, "MOVING TEMPORARY MANIFEST TO MAIN MANIFEST FAILED!\n"+
+		"It is possible that the manifest would have been corrupted. You must move "+
+		"the file: %s to: %s in order to "+
+		"fix the backup manifest.", tmpManifest, backupManifest)
+}
+
+func createBackupFile(h UriHandler, uri *url.URL, req *pb.BackupRequest) (io.WriteCloser, error) {
+	if !h.DirExists("./") {
+		if err := h.CreateDir("./"); err != nil {
+			return nil, errors.Wrap(err, "while creating backup file")
+		}
+	}
+	fileName := backupName(req.ReadTs, req.GroupId)
+	dir := fmt.Sprintf(backupPathFmt, req.UnixTs)
+	if err := h.CreateDir(dir); err != nil {
+		return nil, errors.Wrap(err, "while creating backup file")
+	}
+	backupFile := filepath.Join(dir, fileName)
+	w, err := h.CreateFile(backupFile)
+	return w, errors.Wrap(err, "while creating backup file")
+}
+
+func Load(h UriHandler, uri *url.URL, backupId string, backupNum uint64, fn loadFn) LoadResult {
+	manifests, err := getManifestsToRestore(h, uri, backupId, backupNum)
+	if err != nil {
+		return LoadResult{Err: errors.Wrapf(err, "cannot retrieve manifests")}
+	}
+
+	// Process each manifest, first check that they are valid and then confirm the
+	// backup files for each group exist. Each group in manifest must have a backup file,
+	// otherwise this is a failure and the user must remedy.
+	var since uint64
+	var maxUid, maxNsId uint64
+	for i, manifest := range manifests {
+		if manifest.Since == 0 || len(manifest.Groups) == 0 {
+			continue
+		}
+
+		path := manifests[i].Path
+		for gid := range manifest.Groups {
+			file := filepath.Join(path, backupName(manifest.Since, gid))
+			reader, err := h.Stream(file)
+			if err != nil {
+				return LoadResult{Err: errors.Wrapf(err, "Failed to open %q", file)}
+			}
+			defer reader.Close()
+
+			// Only restore the predicates that were assigned to this group at the time
+			// of the last backup.
+			predSet := manifests[len(manifests)-1].getPredsInGroup(gid)
+
+			groupMaxUid, groupMaxNsId, err := fn(gid,
+				&loadBackupInput{r: reader, preds: predSet, dropOperations: manifest.DropOperations,
+					isOld: manifest.Version == 0})
+			if err != nil {
+				return LoadResult{Err: err}
+			}
+			maxUid = x.Max(maxUid, groupMaxUid)
+			maxNsId = x.Max(maxNsId, groupMaxNsId)
+		}
+		since = manifest.Since
+	}
+
+	return LoadResult{Version: since, MaxLeaseUid: maxUid, MaxLeaseNsId: maxNsId}
+}
+
+// verifyRequest verifies that the manifest satisfies the requirements to process the given
+// restore request.
+func verifyRequest(h UriHandler, uri *url.URL, req *pb.RestoreRequest,
+	currentGroups []uint32) error {
+
+	manifests, err := getManifestsToRestore(h, uri, req.GetBackupId(), req.GetBackupNum())
+	if err != nil {
+		return errors.Wrapf(err, "while retrieving manifests")
+	}
+	if len(manifests) == 0 {
+		return errors.Errorf("No backups with the specified backup ID %s", req.GetBackupId())
+	}
+
+	// TODO(Ahsan): Do we need to verify the manifests again here?
+	if err := verifyManifests(manifests); err != nil {
+		return err
+	}
+
+	lastManifest := manifests[len(manifests)-1]
+	if len(currentGroups) != len(lastManifest.Groups) {
+		return errors.Errorf("groups in cluster and latest backup manifest differ")
+	}
+
+	for _, group := range currentGroups {
+		if _, ok := lastManifest.Groups[group]; !ok {
+			return errors.Errorf("groups in cluster and latest backup manifest differ")
+		}
+	}
+	return nil
+}
+
+func getManifestsToRestore(h UriHandler, uri *url.URL, backupId string,
+	backupNum uint64) ([]*Manifest, error) {
+	if !h.DirExists("") {
+		return nil, errors.Errorf("getManifestsToRestore: The uri path: %q doesn't exists",
+			uri.Path)
+	}
+
+	manifest, err := getConsolidatedManifest(h, uri)
+	if err != nil {
+		return manifest.Manifests, errors.Wrap(err, "Failed to get consolidated manifest: ")
+	}
+	return getFilteredManifests(h, manifest.Manifests, backupId, backupNum)
 }
 
 // loadFn is a function that will receive the current file being read.
@@ -146,13 +297,12 @@ func LoadBackup(location, backupId string, backupNum uint64, creds *x.MinioCrede
 	if err != nil {
 		return LoadResult{Err: err}
 	}
-
 	h, err := NewUriHandler(uri, creds)
 	if err != nil {
 		return LoadResult{Err: errors.Errorf("Unsupported URI: %v", uri)}
 	}
 
-	return h.Load(uri, backupId, backupNum, fn)
+	return Load(h, uri, backupId, backupNum, fn)
 }
 
 // VerifyBackup will access the backup location and verify that the specified backup can
@@ -168,7 +318,7 @@ func VerifyBackup(req *pb.RestoreRequest, creds *x.MinioCredentials, currentGrou
 		return errors.Wrap(err, "VerifyBackup")
 	}
 
-	return h.Verify(uri, req, currentGroups)
+	return verifyRequest(h, uri, req, currentGroups)
 }
 
 // ListBackupManifests scans location l for backup files and returns the list of manifests.
@@ -183,7 +333,7 @@ func ListBackupManifests(l string, creds *x.MinioCredentials) ([]*Manifest, erro
 		return nil, errors.Wrap(err, "ListBackupManifests")
 	}
 
-	m, err := h.GetManifest(uri)
+	m, err := getManifest(h, uri)
 	if err != nil {
 		return nil, err
 	}
@@ -254,34 +404,25 @@ func backupName(since uint64, groupId uint32) string {
 	return fmt.Sprintf(backupNameFmt, since, groupId)
 }
 
-// verifyRequest verifies the manifests satisfy the requirements to process the given
-// restore request.
-func verifyRequest(req *pb.RestoreRequest, manifests []*Manifest, currentGroups []uint32) error {
-	if len(manifests) == 0 {
-		return errors.Errorf("No backups with the specified backup ID %s", req.GetBackupId())
-	}
-
-	if err := verifyManifests(manifests); err != nil {
-		return err
-	}
-
-	lastManifest := manifests[len(manifests)-1]
-	if len(currentGroups) != len(lastManifest.Groups) {
-		return errors.Errorf("groups in cluster and latest backup manifest differ")
-	}
-
-	for _, group := range currentGroups {
-		if _, ok := lastManifest.Groups[group]; !ok {
-			return errors.Errorf("groups in cluster and latest backup manifest differ")
-		}
-	}
-	return nil
-}
-
-func getManifests(manifests []*Manifest, backupId string,
+func getFilteredManifests(h UriHandler, manifests []*Manifest, backupId string,
 	backupNum uint64) ([]*Manifest, error) {
 
-	manifests, err := filterManifests(manifests, backupId)
+	// validManifests are the ones for which the corresponding backup files exists.
+	var validManifests []*Manifest
+	for _, m := range manifests {
+		missingFiles := false
+		for g, _ := range m.Groups {
+			path := filepath.Join(m.Path, backupName(m.Since, g))
+			if !h.FileExists(path) {
+				missingFiles = true
+				break
+			}
+		}
+		if !missingFiles {
+			validManifests = append(validManifests, m)
+		}
+	}
+	manifests, err := filterManifests(validManifests, backupId)
 	if err != nil {
 		return nil, err
 	}
@@ -300,4 +441,336 @@ func getManifests(manifests []*Manifest, backupId string,
 		manifests = manifests[:backupNum]
 	}
 	return manifests, nil
+}
+
+// UriHandler interface is implemented by URI scheme handlers.
+// When adding new scheme handles, for example 'azure://', an object will implement
+// this interface to supply Dgraph with a way to create or load backup files into DB.
+// For all methods below, the URL object is parsed as described in `newHandler' and
+// the Processor object has the DB, estimated tablets size, and backup parameters.
+type UriHandler interface {
+	// CreateDir creates a directory relative to the root path of the handler.
+	CreateDir(path string) error
+	// CreateFile creates a file relative to the root path of the handler. It also makes the
+	// handler's descriptor to point to this file.
+	CreateFile(path string) (io.WriteCloser, error)
+	// DirExists returns true if the directory relative to the root path of the handler exists.
+	DirExists(path string) bool
+	// FileExists returns true if the file relative to the root path of the handler exists.
+	FileExists(path string) bool
+	// JoinPath appends the given path to the root path of the handler.
+	JoinPath(path string) string
+	// ListPaths returns a list of all the valid paths from the given root path. The given root path
+	// should be relative to the handler's root path.
+	ListPaths(path string) []string
+	// Read reads the file at given relative path and returns the read bytes.
+	Read(path string) ([]byte, error)
+	// Rename renames the src file to the destination file.
+	Rename(src, dst string) error
+	// Stream would stream the path via an instance of io.ReadCloser. Close must be called at the
+	// end to release resources appropriately.
+	Stream(path string) (io.ReadCloser, error)
+}
+
+// NewUriHandler parses the requested URI and finds the corresponding UriHandler.
+// If the passed credentials are not nil, they will be used to override the
+// default credentials (only for backups to minio or S3).
+// Target URI formats:
+//   [scheme]://[host]/[path]?[args]
+//   [scheme]:///[path]?[args]
+//   /[path]?[args] (only for local or NFS)
+//
+// Target URI parts:
+//   scheme - service handler, one of: "file", "s3", "minio"
+//     host - remote address. ex: "dgraph.s3.amazonaws.com"
+//     path - directory, bucket or container at target. ex: "/dgraph/backups/"
+//     args - specific arguments that are ok to appear in logs.
+//
+// Global args (if supported by the handler):
+//     secure - true|false turn on/off TLS.
+//      trace - true|false turn on/off HTTP tracing.
+//   compress - true|false turn on/off data compression.
+//    encrypt - true|false turn on/off data encryption.
+//
+// Examples:
+//   s3://dgraph.s3.amazonaws.com/dgraph/backups?secure=true
+//   minio://localhost:9000/dgraph?secure=true
+//   file:///tmp/dgraph/backups
+//   /tmp/dgraph/backups?compress=gzip
+func NewUriHandler(uri *url.URL, creds *x.MinioCredentials) (UriHandler, error) {
+	switch uri.Scheme {
+	case "file", "":
+		return NewFileHandler(uri), nil
+	case "minio", "s3":
+		return NewS3Handler(uri, creds)
+	}
+	return nil, errors.Errorf("Unable to handle url: %s", uri)
+}
+
+// fileHandler is used for 'file:' URI scheme.
+type fileHandler struct {
+	rootDir string
+	prefix  string
+}
+
+func NewFileHandler(uri *url.URL) *fileHandler {
+	h := &fileHandler{}
+	h.rootDir, h.prefix = filepath.Split(uri.Path)
+	return h
+}
+
+func (h *fileHandler) DirExists(path string) bool       { return pathExist(h.JoinPath(path)) }
+func (h *fileHandler) FileExists(path string) bool      { return pathExist(h.JoinPath(path)) }
+func (h *fileHandler) Read(path string) ([]byte, error) { return ioutil.ReadFile(h.JoinPath(path)) }
+
+func (h *fileHandler) JoinPath(path string) string {
+	return filepath.Join(h.rootDir, h.prefix, path)
+}
+func (h *fileHandler) Stream(path string) (io.ReadCloser, error) {
+	return os.Open(h.JoinPath(path))
+}
+func (h *fileHandler) ListPaths(path string) []string {
+	path = h.JoinPath(path)
+	return x.WalkPathFunc(path, func(path string, isDis bool) bool {
+		return true
+	})
+}
+func (h *fileHandler) CreateDir(path string) error {
+	path = h.JoinPath(path)
+	if err := os.MkdirAll(path, 0755); err != nil {
+		return errors.Errorf("Create path failed to create path %s, got error: %v", path, err)
+	}
+	return nil
+}
+
+type fileSyncer struct {
+	fp *os.File
+}
+
+func (fs *fileSyncer) Write(p []byte) (n int, err error) { return fs.fp.Write(p) }
+func (fs *fileSyncer) Close() error {
+	if err := fs.fp.Sync(); err != nil {
+		return errors.Wrapf(err, "while syncing file: %s", fs.fp.Name())
+	}
+	err := fs.fp.Close()
+	return errors.Wrapf(err, "while closing file: %s", fs.fp.Name())
+}
+
+func (h *fileHandler) CreateFile(path string) (io.WriteCloser, error) {
+	path = h.JoinPath(path)
+	fp, err := os.Create(path)
+	return &fileSyncer{fp}, errors.Wrapf(err, "File handler failed to create file %s", path)
+}
+
+func (h *fileHandler) Rename(src, dst string) error {
+	src = h.JoinPath(src)
+	dst = h.JoinPath(dst)
+	return os.Rename(src, dst)
+}
+
+// pathExist checks if a path (file or dir) is found at target.
+// Returns true if found, false otherwise.
+func pathExist(path string) bool {
+	_, err := os.Stat(path)
+	if err == nil {
+		return true
+	}
+	return !os.IsNotExist(err) && !os.IsPermission(err)
+}
+
+// S3 Handler.
+
+// FillRestoreCredentials fills the empty values with the default credentials so that
+// a restore request is sent to all the groups with the same credentials.
+func FillRestoreCredentials(location string, req *pb.RestoreRequest) error {
+	uri, err := url.Parse(location)
+	if err != nil {
+		return err
+	}
+
+	defaultCreds := credentials.Value{
+		AccessKeyID:     req.AccessKey,
+		SecretAccessKey: req.SecretKey,
+		SessionToken:    req.SessionToken,
+	}
+	provider := x.MinioCredentialsProvider(uri.Scheme, defaultCreds)
+
+	creds, _ := provider.Retrieve() // Error is always nil.
+
+	req.AccessKey = creds.AccessKeyID
+	req.SecretKey = creds.SecretAccessKey
+	req.SessionToken = creds.SessionToken
+
+	return nil
+}
+
+// s3Handler is used for 's3:' and 'minio:' URI schemes.
+type s3Handler struct {
+	bucketName, objectPrefix string
+	creds                    *x.MinioCredentials
+	uri                      *url.URL
+	mc                       *x.MinioClient
+}
+
+// NewS3Handler creates a new session, checks valid bucket at uri.Path, and configures a
+// minio client. It also fills in values used by the handler in subsequent calls.
+// Returns a new S3 minio client, otherwise a nil client with an error.
+func NewS3Handler(uri *url.URL, creds *x.MinioCredentials) (*s3Handler, error) {
+	h := &s3Handler{
+		creds: creds,
+		uri:   uri,
+	}
+	mc, err := x.NewMinioClient(uri, creds)
+	if err != nil {
+		return nil, err
+	}
+	h.mc = mc
+	h.bucketName, h.objectPrefix = mc.ParseBucketAndPrefix(uri.Path)
+	return h, nil
+}
+
+func (h *s3Handler) CreateDir(path string) error { return nil }
+func (h *s3Handler) DirExists(path string) bool  { return true }
+
+func (h *s3Handler) FileExists(path string) bool {
+	objectPath := h.getObjectPath(path)
+	_, err := h.mc.StatObject(h.bucketName, objectPath, minio.StatObjectOptions{})
+	if err != nil {
+		errResponse := minio.ToErrorResponse(err)
+		if errResponse.Code == "NoSuchKey" {
+			return false
+		} else {
+			glog.Errorf("Failed to verify object existence: %v", err)
+			return false
+		}
+	}
+	return true
+}
+
+func (h *s3Handler) JoinPath(path string) string {
+	return filepath.Join(h.bucketName, h.objectPrefix, path)
+}
+
+func (h *s3Handler) Read(path string) ([]byte, error) {
+	objectPath := h.getObjectPath(path)
+	var buf bytes.Buffer
+
+	reader, err := h.mc.GetObject(h.bucketName, objectPath, minio.GetObjectOptions{})
+	if err != nil {
+		return buf.Bytes(), errors.Wrap(err, "Failed to read s3 object")
+	}
+	defer reader.Close()
+
+	if _, err := buf.ReadFrom(reader); err != nil {
+		return buf.Bytes(), errors.Wrap(err, "Failed to read the s3 object")
+	}
+	return buf.Bytes(), nil
+}
+
+func (h *s3Handler) Stream(path string) (io.ReadCloser, error) {
+	objectPath := h.getObjectPath(path)
+	reader, err := h.mc.GetObject(h.bucketName, objectPath, minio.GetObjectOptions{})
+	if err != nil {
+		return nil, err
+	}
+	return reader, nil
+}
+
+func (h *s3Handler) ListPaths(path string) []string {
+	var paths []string
+	done := make(chan struct{})
+	defer close(done)
+	path = h.getObjectPath(path)
+	for object := range h.mc.ListObjects(h.bucketName, path, true, done) {
+		paths = append(paths, object.Key)
+	}
+	return paths
+}
+
+type s3Writer struct {
+	pwriter    *io.PipeWriter
+	preader    *io.PipeReader
+	bucketName string
+	cerr       chan error
+}
+
+func (sw *s3Writer) Write(p []byte) (n int, err error) { return sw.pwriter.Write(p) }
+func (sw *s3Writer) Close() error {
+	if sw.pwriter == nil {
+		return nil
+	}
+	if err := sw.pwriter.CloseWithError(nil); err != nil && err != io.EOF {
+		glog.Errorf("Unexpected error when closing pipe: %v", err)
+	}
+	sw.pwriter = nil
+	glog.V(2).Infof("Backup waiting for upload to complete.")
+	return <-sw.cerr
+}
+
+// upload will block until it's done or an error occurs.
+func (sw *s3Writer) upload(mc *x.MinioClient, object string) {
+	f := func() error {
+		start := time.Now()
+
+		// We don't need to have a progress object, because we're using a Pipe. A write to Pipe
+		// would block until it can be fully read. So, the rate of the writes here would be equal to
+		// the rate of upload. We're already tracking progress of the writes in stream.Lists, so no
+		// need to track the progress of read. By definition, it must be the same.
+		//
+		// PutObject would block until sw.preader returns EOF.
+		n, err := mc.PutObject(sw.bucketName, object, sw.preader, -1, minio.PutObjectOptions{})
+		glog.V(2).Infof("Backup sent %d bytes. Time elapsed: %s",
+			n, time.Since(start).Round(time.Second))
+
+		if err != nil {
+			// This should cause Write to fail as well.
+			glog.Errorf("Backup: Closing RW pipe due to error: %v", err)
+			if err := sw.pwriter.Close(); err != nil {
+				return err
+			}
+			if err := sw.preader.Close(); err != nil {
+				return err
+			}
+		}
+		return err
+	}
+	sw.cerr <- f()
+}
+
+func (h *s3Handler) CreateFile(path string) (io.WriteCloser, error) {
+	objectPath := h.getObjectPath(path)
+	glog.V(2).Infof("Sending data to %s blob %q ...", h.uri.Scheme, objectPath)
+
+	sw := &s3Writer{
+		bucketName: h.bucketName,
+		cerr:       make(chan error, 1),
+	}
+	sw.preader, sw.pwriter = io.Pipe()
+	go sw.upload(h.mc, objectPath)
+	return sw, nil
+}
+
+func (h *s3Handler) Rename(srcPath, dstPath string) error {
+	src := minio.NewSourceInfo(h.bucketName, srcPath, nil)
+	dst, err := minio.NewDestinationInfo(h.bucketName, dstPath, nil, nil)
+	if err != nil {
+		return errors.Wrap(err, "Rename failed to create dstInfo")
+	}
+	// We try copying 100 times, if it still fails, then the user should manually rename.
+	err = x.RetryUntilSuccess(100, time.Second, func() error {
+		if err := h.mc.CopyObject(dst, src); err != nil {
+			return errors.Wrapf(err, "While renaming object in s3, copy failed ")
+		}
+		return nil
+	})
+	if err != nil {
+		return err
+	}
+
+	err = h.mc.RemoveObject(h.bucketName, srcPath)
+	return errors.Wrap(err, "Rename failed to remove temporary file")
+}
+
+func (h *s3Handler) getObjectPath(path string) string {
+	return filepath.Join(h.objectPrefix, path)
 }
